PSMDB-971 Implement KMIP encryption
This commit implements encryption using a master key stored on a KMIP
server. The server is required to implement the LOCATE, REGISTER and GET
operations.

Configuration parameters work similarly to MongoDB enterprise, except:
* since the used KMIP library requires client information in a separate
  CA and Key file, two parameters are required for that. PSMDB has a new
  "clientKeyFile" parameter
* there is no automation for key name generation, the keyIdentifier
  parameter is required

Implementation is based on our fork of the KMIP library, and its C++
bindings. The library itself isn't included in the git repository or in
the build process. The build process expect to find an already built
library in the include/lib path.

To build the KMIP library:

* git checkout https://github.com/dutow/libkmip.git
* cd libkmip
* git checkout locate-support
* cmake -DCMAKE_BUILD_TYPE=...
  -DCMAKE_INSTALL_PREFIX=/where/to/install/it