PSMDB-1148 Vault: read a token internally
Clients of the `vaultReadKey` and the `vaultWriteKey` functions
don't need to read a Vault token themselves anymore.
The patch also eliminates code duplication, reduces lifetime of the
read token, securely destroys the token's memory. On top of that,
`perconadecrypt` now checks the permissions of the encryption key
file and the token file for being too open.