PSMDB-1228: Repair the build for the older platforms
Not all the platforms Percona Server for MongoDB is currently
supported on implement the `std::filesystem::permissions` function
in their `libstdc++` library.
Additionally, on some older platforms the output operator for
`nullptr` is ambiguous in the unit tests.

The build was broken in the commit
b988c5193627296fb52b4494235b86c5a7e5f88a