PSMDB-1447 Fix `db_auth_test`
The `AuthorizationSessionTest/ExpirationWithSecurityTokenNOK` test case
doesn't pass due to the `startBackup` action being listed in the
`dbAdmin` role and not being listed in the actions valid on a database
for the serverless configuration. (@see
`src/mongo/db/auth/action_types.yml`, line 295).

Enabling `startBackup` for serverless requires further research on its
potential implications. For now, we allow the `dbAdmin` role to execute
the `startBackup` action only for the single tenant configuration.