PSMDB-1447 Fix `mongo_embedded_test`
The following commands:

- `_dropConnectionsToMongot`,
- `_mongotConnPoolStats`,
- `createSearchIndexes`,
- `dropSearchIndex`,
- `listSearchIndexes`,
- `updateSearchIndex`

are part of the `db/commands/search_index_commands` library, which
is a dependency of the `db/commands/standalone` library. The latter
comprises the commands that should be available in both `mongod`
and `embedded` (@see file `src/mongo/db/commands/SConscript`,
definition of the `standalone` library on line 457).